ref: move handlePublishListItems to controllers.publish
